third_party/WebKit/LayoutTests/media/track/track-load-from-src-readyState.html

<!DOCTYPE html>
<title>Tests the load event on HTMLTrackElement and LOADED readyState on TextTrack when src is set from JavaScript.</title>
<video>
    <track>
</video>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
async_test(function(t) {
    var track = document.querySelector("track");
    assert_equals(track.readyState, HTMLTrackElement.NONE);

    track.onload = t.step_func_done(function () {
        assert_equals(track.readyState, HTMLTrackElement.LOADED);
    });

    track.src = "captions-webvtt/tc004-webvtt-file.vtt";
    track.track.mode = "hidden";
});
</script>
